+++ packages/features/ee/workflows/components/WorkflowDetailsPage.tsx
@@ -108,6 +108,8 @@ export default function WorkflowDetailsPage(props: Props) {
     form.setValue("steps", steps);
   };
 
+  const additionalComponent = { MultiValue };
+
   return (
     <>
       <div className="my-8 sm:my-0 md:flex">
@@ -140,6 +142,7 @@ export default function WorkflowDetailsPage(props: Props) {
                   closeMenuOnSelect={false}
                   hideSelectedOptions={false}
                   isLoading={isLoading}
+                  components={additionalComponent}
                 />
               );
             }}
@@ -214,3 +217,9 @@ export default function WorkflowDetailsPage(props: Props) {
     </>
   );
 }
+
+const MultiValue = ({ index, getValue }: { index: number; getValue: any }) => {
+  const { t } = useLocale();
+
+  return <>{!index && <div>{t("nr_event_type", { count: getValue().length })}</div>}</>;
+};
