+++ packages/ui/components/form/select/Select.tsx
@@ -1,26 +1,21 @@
 import { useId } from "@radix-ui/react-id";
 import * as React from "react";
-import type { GroupBase, Props, SingleValue, MultiValue, OptionProps } from "react-select";
+import type { GroupBase, Props, SingleValue, MultiValue } from "react-select";
 import ReactSelect from "react-select";
+import type { SelectComponents } from "react-select/dist/declarations/src/components";
 
 import cx from "@calcom/lib/classNames";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 
 import { Label } from "../inputs/Label";
+import { CheckboxComponent } from "./components";
 import { getReactSelectProps } from "./selectTheme";
-import { SelectComponents } from "react-select/dist/declarations/src/components";
 
 export type SelectProps<
   Option,
   IsMulti extends boolean = false,
   Group extends GroupBase<Option> = GroupBase<Option>
 > = Props<Option, IsMulti, Group> & { variant?: "default" | "checkbox" };
-
-type InputOptionProps = 
-{
-  Option?: React.ComponentType<OptionProps<Option, boolean, GroupBase<Option>>>;
-  inputProps: OptionProps<Option, boolean, GroupBase<Option>>;
-};
 
 export const Select = <
   Option,
@@ -34,21 +29,21 @@ export const Select = <
   ...props
 }: SelectProps<Option, IsMulti, Group>) => {
   const { classNames, className, ...restProps } = props;
-  const additonalComponents = { MultiValue };
   const isMultiSelectCheckbox = variant === "checkbox" && isMulti;
 
-  const componentProps = isMultiSelectCheckbox ? {
-    ...additonalComponents,
-    Option: (props: InputOptionProps["inputProps"]) => (
-      <InputOption inputProps={props} Option={components?.Option as InputOptionProps["Option"]} />
-    )
-  } : components
   const reactSelectProps = React.useMemo(() => {
+    const componentProps = isMultiSelectCheckbox
+      ? {
+          ...components,
+          Option: CheckboxComponent,
+        }
+      : components;
+
     return getReactSelectProps<Option, IsMulti, Group>({
-      components: componentProps as Partial<SelectComponents<Option, IsMulti, Group>> || {},
+      components: (componentProps as Partial<SelectComponents<Option, IsMulti, Group>>) || {},
       menuPlacement,
     });
-  }, [components, menuPlacement]);
+  }, [components, menuPlacement, isMultiSelectCheckbox]);
 
   // Annoyingly if we update styles here we have to update timezone select too
   // We cant create a generate function for this as we can't force state changes - onSelect styles dont change for example
@@ -68,7 +63,7 @@ export const Select = <
         dropdownIndicator: () => "text-default",
         control: (state) =>
           cx(
-            "bg-default border-default !min-h-9 h-9 text-sm leading-4 placeholder:text-sm placeholder:font-normal focus-within:ring-2 focus-within:ring-emphasis hover:border-emphasis rounded-md border ",
+            "bg-default border-default !min-h-9 text-sm leading-4 placeholder:text-sm placeholder:font-normal focus-within:ring-2 focus-within:ring-emphasis hover:border-emphasis rounded-md border ",
             state.isMulti
               ? variant === "checkbox"
                 ? "px-3 py-2 h-fit"
@@ -213,41 +208,3 @@ export function SelectWithValidation<
     </div>
   );
 }
-
-export type Option = {
-  value: string;
-  label: string;
-};
-
-const InputOption: React.FC<InputOptionProps> = ({
-  Option,
-  inputProps
-}) => {
-  const { isDisabled, isFocused, isSelected, children, innerProps, className, ...rest } = inputProps;
-  const props = {
-    ...innerProps,
-  };
-
-  return (
-    Option ? <Option
-      {...rest}
-      isDisabled={isDisabled}
-      isFocused={isFocused}
-      isSelected={isSelected}
-      innerProps={props}>
-      <input
-        type="checkbox"
-        className="text-primary-600 focus:ring-primary-500 border-default h-4 w-4 rounded ltr:mr-2 rtl:ml-2"
-        checked={isSelected}
-        readOnly
-      />
-      {children}
-    </Option> : <></>
-  );
-};
-
-const MultiValue = ({ index, getValue }: { index: number; getValue: any }) => {
-  const { t } = useLocale();
-
-  return <>{!index && <div>{t("nr_event_type", { count: getValue().length })}</div>}</>;
-};
