+++ packages/ui/components/form/select/components.tsx
@@ -67,3 +67,38 @@ export const IconLeading = ({ icon, children, ...props }: IconLeadingProps) => {
     </reactSelectComponents.Control>
   );
 };
+
+type Option = {
+  value: string;
+  label: string;
+};
+
+export const CheckboxComponent: React.FC<OptionProps<Option, boolean, GroupBase<any>>> = ({
+  isDisabled,
+  isFocused,
+  isSelected,
+  children,
+  innerProps,
+  ...rest
+}) => {
+  const props = {
+    ...innerProps,
+  };
+
+  return (
+    <reactSelectComponents.Option
+      {...rest}
+      isDisabled={isDisabled}
+      isFocused={isFocused}
+      isSelected={isSelected}
+      innerProps={props}>
+      <input
+        type="checkbox"
+        className="text-primary-600 focus:ring-primary-500 border-default h-4 w-4 rounded ltr:mr-2 rtl:ml-2"
+        checked={isSelected}
+        readOnly
+      />
+      {children}
+    </reactSelectComponents.Option>
+  );
+};
